+++ crest_data_systems_microsoft_defender/CHANGELOG.md
@@ -1,5 +1,9 @@
-# CHANGELOG - Microsoft Defender
+# CHANGELOG - Microsoft 365 Defender
 
-## 1.0.0
+## 1.0.1 / 2022-08-10
+
+* [Added] Pagination of API to fetch records more than 10,000.
+
+## 1.0.0 / 2022-07-05
 
 * [Added] Initial release.